DAG-main/MAS-main/financial_intelligence/groq-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.10.0
tavily-python==0.3.0
python-dotenv==1.0.0
pydantic==2.5.3